ArbolGenealogicoService
Layout de coordendas calculas perfectamente

+++ ArbolGenealogicoWPF/ArbolGenealogicoWPF/GridCoord.cs
@@ -3,6 +3,6 @@ namespace ArbolGenealogicoWPF
     public class GridCoord
     {
         public int Row { get; set; }   // Nivel generacional (diferencia de nivel padre-hijo)
-        public int Col { get; set; }   // Posición horizontal (relación de igual hermanos y pareja)
+        public double Col { get; set; }   // Posición horizontal (relación de igual hermanos y pareja)
     }
 }
